cmd/utils: fix NetworkId default when -dev is set (#16833)
Prior to this change, when geth was started with `geth -dev -rpc`,
it would report a network id of `1` in response to the `net_version` RPC
request. But the actual network id it used to verify transactions
was `200666`.

This change causes geth instead respond with `200666` to the `net_version`
RPC when geth is started with `geth -dev -rpc`.